api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
dnspython==2.5.0
httpx==0.26.0
python-multipart==0.0.6
pytest==7.4.4
pytest-asyncio==0.23.3
google-generativeai==0.3.2
